REFACTOR(MainActivity.java -> DonateActivity.java), FIX(recycleview items): Put in manifest file MainMenuActivity to be first and fix to not show previously clicked exercises in recycleview.
