+++ app/src/main/AndroidManifest.xml
@@ -15,11 +15,11 @@
             android:name=".ExercisesActivity"
             android:exported="false" />
         <activity
-            android:name=".MainMenuActivity"
-            android:exported="false" />
+            android:name=".DonateActivity"
+            android:exported="false"/>
         <activity
-            android:name=".MainActivity"
-            android:exported="true">
+            android:name=".MainMenuActivity"
+            android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
